renamed modul texas til token-klient
