+++ token-klient/src/main/kotlin/no/nav/familie/felles/tokenklient/TexasTokenResponse.kt
@@ -1,4 +1,4 @@
-package no.nav.familie.felles.texas
+package no.nav.familie.felles.tokenklient
 
 import com.fasterxml.jackson.annotation.JsonProperty
