+++ token-klient/src/test/kotlin/no/nav/familie/felles/tokenklient/TexasRestClientFactoryTest.kt
@@ -1,4 +1,4 @@
-package no.nav.familie.felles.texas
+package no.nav.familie.felles.tokenklient
 
 import io.mockk.mockk
 import no.nav.familie.log.interceptor.ConsumerIdClientInterceptor
